Project1-main/django_backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.1
python-dotenv==1.0.0
django-filter==23.5
pandas==2.1.4
bcrypt==4.1.2
python-decouple==3.8
Pillow==10.1.0
tzdata==2024.1